[Documentation/Unity ID] Made some major fixes.
The documentation site is now nearly complete. Backwards compatibility
with older documentation versions has been added, which means things like
the code language setting (that lets you change between C# and JS) is now
supported, along with the history table available in documentation versions
5.0.x to 5.2.x. Also made some major fixes to the theming of the Unity ID
platform.

The theme version was bumped to 1.1 as well, with the codename of Nitrogen.

Signed-off-by: Kamran Mackey <kamranm1200@gmail.com>